Create FeelingWords-FullList.txt
Purchased list of words limited to emotional content.  This list is sorted alphabetically, except near the end where they are sorted by when the author added them to the list.  Work needs to be done to make the list parsable, or translated into a format where they can be scored on various emotional matrices.